chore: Refactor main.ts to use unified configuration for request and frontend
