+++ src/main.ts
@@ -2,10 +2,6 @@
 
 import { ConfigUtil } from "./utils/ConfigUtil";
 import { RequestConfig, FrontendGeneratorConfig } from "./interfaces";
-import * as fs from 'fs-extra';
-import path from "path";
-import { exec } from "child_process";
-import * as prettier from "prettier";
 
 import { ComponentGenerator } from "./component-generator";
 import { StylesGenerator } from "./styles-generator";
@@ -32,62 +28,8 @@ console.log(`App: ${frontendConfig.app}`);
 console.log(`Output Directory: ${frontendConfig.outputDir}`);
 console.log(`Components: ${frontendConfig.components.join(", ")}`);
 
-async function copyStaticFiles(destDir: string) {
-    try {
-        const staticPath = path.resolve(__dirname, '../static');
-        await fs.copy(staticPath, destDir, {
-            overwrite: true,
-        });
-        console.log('Arquivos estáticos copiados com sucesso.');
-    } catch (err) {
-        console.error('Erro ao copiar arquivos estáticos:', err);
-    }
-}
-
-async function formatFiles(destDir: string) {
-    try {
-        const configFile = await prettier.resolveConfigFile(destDir);
-
-        const options = {
-            config: configFile,
-            ignorePath: path.join(destDir, '.prettierignore'),
-            editorconfig: true,
-        };
-
-        const files = await fs.promises.readdir(destDir);
-
-        for (const filePath of files) {
-            const content = await fs.promises.readFile(path.join(destDir, filePath), 'utf8');
-            const formatted = await prettier.format(content, { ...options, filepath: filePath });
-            await fs.promises.writeFile(path.join(destDir, filePath), formatted);
-        }
-
-        console.log('Arquivos gerados formatados com sucesso.');
-    } catch (err) {
-        console.error('Erro ao formatar arquivos gerados:', err);
-    }
-}
-
-async function runNpmInstall(directory: string): Promise<void> {
-    console.log('Instalando dependências via npm install...');
-    return new Promise((resolve, reject) => {
-        exec('npm install', { cwd: directory }, (error, stdout, stderr) => {
-            if (error) {
-                console.error(`Erro ao executar npm install: ${error.message}`);
-                reject(error);
-            }
-            if (stderr) {
-                console.error(`stderr: ${stderr}`);
-            }
-            console.log(`stdout: ${stdout}`);
-            console.log('Dependências instaladas com sucesso.');
-            resolve();
-        });
-    });
-}
-
 async function main() {
-    await copyStaticFiles(frontendConfig.outputDir);
+    await ConfigUtil.copyStaticFiles(frontendConfig.outputDir);
 
     const generators: { [key: string]: any } = {
         "component": ComponentGenerator,
@@ -112,8 +54,8 @@ async function main() {
         }
     }
 
-    await runNpmInstall(frontendConfig.outputDir);
-    await formatFiles(frontendConfig.outputDir);
+    await ConfigUtil.runNpmInstall(frontendConfig.outputDir);
+    await ConfigUtil.formatFiles(frontendConfig.outputDir);
 }
 
 main();
